requirements.txt
flask==3.1.0
python-dotenv==1.0.0
pyodbc==5.2.0
pypdf2==3.0.1
pdfplumber==0.12.0
reportlab==4.4.2
anthropic==0.57.1 